+++ src/http/http_endpoint.h
@@ -5,6 +5,7 @@
 #include "ccf/ds/logger.h"
 #include "enclave/client_endpoint.h"
 #include "enclave/rpc_map.h"
+#include "enclave/rpc_sessions.h"
 #include "http_parser.h"
 #include "http_rpc_context.h"
 
@@ -14,18 +15,28 @@
 
 namespace http
 {
-  class HTTPEndpoint : public enclave::TLSEndpoint
+  class ErrorReporter
+  {
+  public:
+    virtual ~ErrorReporter() {}
+    virtual void report_parsing_error(tls::ConnID) = 0;
+  };
+
+  class HTTPEndpoint : public ccf::TLSEndpoint
   {
   protected:
     http::Parser& p;
+    std::shared_ptr<ErrorReporter> error_reporter;
 
     HTTPEndpoint(
       http::Parser& p_,
-      int64_t session_id,
+      tls::ConnID session_id,
       ringbuffer::AbstractWriterFactory& writer_factory,
-      std::unique_ptr<tls::Context> ctx) :
+      std::unique_ptr<tls::Context> ctx,
+      const std::shared_ptr<ErrorReporter>& error_reporter = nullptr) :
       TLSEndpoint(session_id, writer_factory, std::move(ctx)),
-      p(p_)
+      p(p_),
+      error_reporter(error_reporter)
     {}
 
   public:
@@ -74,7 +85,10 @@ namespace http
         }
         catch (const std::exception& e)
         {
-          LOG_FAIL_FMT("Error parsing HTTP request");
+          if (error_reporter)
+          {
+            error_reporter->report_parsing_error(session_id);
+          }
           LOG_DEBUG_FMT("Error parsing HTTP request: {}", e.what());
 
           auto response = http::Response(HTTP_STATUS_BAD_REQUEST);
@@ -84,7 +98,7 @@ namespace http
           // bytes. Instead insert it at the end of this message, verbatim
           auto body_s = fmt::format(
             "Unable to parse data as a HTTP request. Error message is: {}\n"
-            "Error occured while parsing fragment:\n",
+            "Error occurred while parsing fragment:\n",
             e.what());
           std::vector<uint8_t> response_body(
             std::begin(body_s), std::end(body_s));
@@ -104,21 +118,26 @@ namespace http
   private:
     http::RequestParser request_parser;
 
-    std::shared_ptr<enclave::RPCMap> rpc_map;
-    std::shared_ptr<enclave::RpcHandler> handler;
-    std::shared_ptr<enclave::SessionContext> session_ctx;
-    int64_t session_id;
-    enclave::ListenInterfaceID interface_id;
-    size_t request_index = 0;
+    std::shared_ptr<ccf::RPCMap> rpc_map;
+    std::shared_ptr<ccf::RpcHandler> handler;
+    std::shared_ptr<ccf::SessionContext> session_ctx;
+    tls::ConnID session_id;
+    ccf::ListenInterfaceID interface_id;
 
   public:
     HTTPServerEndpoint(
-      std::shared_ptr<enclave::RPCMap> rpc_map,
-      int64_t session_id,
-      const enclave::ListenInterfaceID& interface_id,
+      std::shared_ptr<ccf::RPCMap> rpc_map,
+      tls::ConnID session_id,
+      const ccf::ListenInterfaceID& interface_id,
       ringbuffer::AbstractWriterFactory& writer_factory,
-      std::unique_ptr<tls::Context> ctx) :
-      HTTPEndpoint(request_parser, session_id, writer_factory, std::move(ctx)),
+      std::unique_ptr<tls::Context> ctx,
+      const std::shared_ptr<ErrorReporter>& error_reporter = nullptr) :
+      HTTPEndpoint(
+        request_parser,
+        session_id,
+        writer_factory,
+        std::move(ctx),
+        error_reporter),
       request_parser(*this),
       rpc_map(rpc_map),
       session_id(session_id),
@@ -146,20 +165,15 @@ namespace http
       {
         if (session_ctx == nullptr)
         {
-          session_ctx = std::make_shared<enclave::SessionContext>(
+          session_ctx = std::make_shared<ccf::SessionContext>(
             session_id, peer_cert(), interface_id);
         }
 
-        std::shared_ptr<enclave::RpcContext> rpc_ctx = nullptr;
+        std::shared_ptr<http::HttpRpcContext> rpc_ctx = nullptr;
         try
         {
           rpc_ctx = std::make_shared<HttpRpcContext>(
-            request_index++,
-            session_ctx,
-            verb,
-            url,
-            std::move(headers),
-            std::move(body));
+            session_ctx, verb, url, std::move(headers), std::move(body));
         }
         catch (std::exception& e)
         {
@@ -227,7 +241,7 @@ namespace http
   };
 
   class HTTPClientEndpoint : public HTTPEndpoint,
-                             public enclave::ClientEndpoint,
+                             public ccf::ClientEndpoint,
                              public http::ResponseProcessor
   {
   private:
@@ -235,7 +249,7 @@ namespace http
 
   public:
     HTTPClientEndpoint(
-      int64_t session_id,
+      tls::ConnID session_id,
       ringbuffer::AbstractWriterFactory& writer_factory,
       std::unique_ptr<tls::Context> ctx) :
       HTTPEndpoint(response_parser, session_id, writer_factory, std::move(ctx)),