+++ src/node/rpc/test/frontend_test.cpp
@@ -313,7 +313,7 @@ public:
   void record_ctx(ccf::endpoints::EndpointContext& ctx)
   {
     last_caller_cert =
-      crypto::cert_der_to_pem(ctx.rpc_ctx->session->caller_cert);
+      crypto::cert_der_to_pem(ctx.rpc_ctx->get_session_context()->caller_cert);
     if (const auto uci = ctx.try_get_caller<UserCertAuthnIdentity>())
     {
       last_caller_id = uci->user_id;
@@ -469,16 +469,16 @@ auto invalid_caller_der = crypto::make_verifier(invalid_caller) -> cert_der();
 
 auto anonymous_caller_der = std::vector<uint8_t>();
 
-auto user_session = make_shared<enclave::SessionContext>(
-  enclave::InvalidSessionId, user_caller_der);
-auto backup_user_session = make_shared<enclave::SessionContext>(
-  enclave::InvalidSessionId, user_caller_der);
-auto invalid_session = make_shared<enclave::SessionContext>(
-  enclave::InvalidSessionId, invalid_caller_der);
-auto member_session = make_shared<enclave::SessionContext>(
-  enclave::InvalidSessionId, member_caller_der);
-auto anonymous_session = make_shared<enclave::SessionContext>(
-  enclave::InvalidSessionId, anonymous_caller_der);
+auto user_session =
+  make_shared<ccf::SessionContext>(ccf::InvalidSessionId, user_caller_der);
+auto backup_user_session =
+  make_shared<ccf::SessionContext>(ccf::InvalidSessionId, user_caller_der);
+auto invalid_session =
+  make_shared<ccf::SessionContext>(ccf::InvalidSessionId, invalid_caller_der);
+auto member_session =
+  make_shared<ccf::SessionContext>(ccf::InvalidSessionId, member_caller_der);
+auto anonymous_session =
+  make_shared<ccf::SessionContext>(ccf::InvalidSessionId, anonymous_caller_der);
 
 UserId user_id;
 
@@ -543,7 +543,7 @@ TEST_CASE("process with signatures")
     constexpr auto rpc_name = "/this_rpc_doesnt_exist";
     const auto invalid_call = create_simple_request(rpc_name);
     const auto serialized_call = invalid_call.build_request();
-    auto rpc_ctx = enclave::make_rpc_context(user_session, serialized_call);
+    auto rpc_ctx = ccf::make_rpc_context(user_session, serialized_call);
 
     const auto serialized_response = frontend.process(rpc_ctx).value();
     auto response = parse_response(serialized_response);
@@ -558,9 +558,9 @@ TEST_CASE("process with signatures")
     const auto serialized_signed_call = signed_call.build_request();
 
     auto simple_rpc_ctx =
-      enclave::make_rpc_context(user_session, serialized_simple_call);
+      ccf::make_rpc_context(user_session, serialized_simple_call);
     auto signed_rpc_ctx =
-      enclave::make_rpc_context(user_session, serialized_signed_call);
+      ccf::make_rpc_context(user_session, serialized_signed_call);
 
     INFO("Unsigned RPC");
     {
@@ -585,9 +585,9 @@ TEST_CASE("process with signatures")
     const auto serialized_signed_call = signed_call.build_request();
 
     auto simple_rpc_ctx =
-      enclave::make_rpc_context(user_session, serialized_simple_call);
+      ccf::make_rpc_context(user_session, serialized_simple_call);
     auto signed_rpc_ctx =
-      enclave::make_rpc_context(user_session, serialized_signed_call);
+      ccf::make_rpc_context(user_session, serialized_signed_call);
 
     INFO("Unsigned RPC");
     {
@@ -619,11 +619,11 @@ TEST_CASE("process with caller")
     const auto simple_call = create_simple_request("/empty_function_no_auth");
     const auto serialized_simple_call = simple_call.build_request();
     auto authenticated_rpc_ctx =
-      enclave::make_rpc_context(user_session, serialized_simple_call);
+      ccf::make_rpc_context(user_session, serialized_simple_call);
     auto invalid_rpc_ctx =
-      enclave::make_rpc_context(invalid_session, serialized_simple_call);
+      ccf::make_rpc_context(invalid_session, serialized_simple_call);
     auto anonymous_rpc_ctx =
-      enclave::make_rpc_context(anonymous_session, serialized_simple_call);
+      ccf::make_rpc_context(anonymous_session, serialized_simple_call);
 
     INFO("Valid authentication");
     {
@@ -658,11 +658,11 @@ TEST_CASE("process with caller")
     const auto simple_call = create_simple_request("/empty_function");
     const auto serialized_simple_call = simple_call.build_request();
     auto authenticated_rpc_ctx =
-      enclave::make_rpc_context(user_session, serialized_simple_call);
+      ccf::make_rpc_context(user_session, serialized_simple_call);
     auto invalid_rpc_ctx =
-      enclave::make_rpc_context(invalid_session, serialized_simple_call);
+      ccf::make_rpc_context(invalid_session, serialized_simple_call);
     auto anonymous_rpc_ctx =
-      enclave::make_rpc_context(anonymous_session, serialized_simple_call);
+      ccf::make_rpc_context(anonymous_session, serialized_simple_call);
 
     INFO("Valid authentication");
     {
@@ -708,7 +708,7 @@ TEST_CASE("No certs table")
 
   INFO("Authenticated caller");
   {
-    auto rpc_ctx = enclave::make_rpc_context(user_session, serialized_call);
+    auto rpc_ctx = ccf::make_rpc_context(user_session, serialized_call);
     std::vector<uint8_t> serialized_response =
       frontend.process(rpc_ctx).value();
     auto response = parse_response(serialized_response);
@@ -717,8 +717,7 @@ TEST_CASE("No certs table")
 
   INFO("Anonymous caller");
   {
-    auto rpc_ctx =
-      enclave::make_rpc_context(anonymous_session, serialized_call);
+    auto rpc_ctx = ccf::make_rpc_context(anonymous_session, serialized_call);
     std::vector<uint8_t> serialized_response =
       frontend.process(rpc_ctx).value();
     auto response = parse_response(serialized_response);
@@ -741,7 +740,7 @@ TEST_CASE("Member caller")
   SUBCASE("valid caller")
   {
     auto member_rpc_ctx =
-      enclave::make_rpc_context(member_session, serialized_call);
+      ccf::make_rpc_context(member_session, serialized_call);
     std::vector<uint8_t> serialized_response =
       frontend.process(member_rpc_ctx).value();
     auto response = parse_response(serialized_response);
@@ -750,7 +749,7 @@ TEST_CASE("Member caller")
 
   SUBCASE("invalid caller")
   {
-    auto rpc_ctx = enclave::make_rpc_context(user_session, serialized_call);
+    auto rpc_ctx = ccf::make_rpc_context(user_session, serialized_call);
     std::vector<uint8_t> serialized_response =
       frontend.process(rpc_ctx).value();
     auto response = parse_response(serialized_response);
@@ -774,7 +773,7 @@ TEST_CASE("JsonWrappedEndpointFunction")
       echo_call.set_body(serialized_body.data(), serialized_body.size());
       const auto serialized_call = echo_call.build_request();
 
-      auto rpc_ctx = enclave::make_rpc_context(user_session, serialized_call);
+      auto rpc_ctx = ccf::make_rpc_context(user_session, serialized_call);
       auto response = parse_response(frontend.process(rpc_ctx).value());
       CHECK(response.status == HTTP_STATUS_OK);
 
@@ -797,7 +796,7 @@ TEST_CASE("JsonWrappedEndpointFunction")
 
       const auto serialized_call = echo_call.build_request();
 
-      auto rpc_ctx = enclave::make_rpc_context(user_session, serialized_call);
+      auto rpc_ctx = ccf::make_rpc_context(user_session, serialized_call);
       auto response = parse_response(frontend.process(rpc_ctx).value());
       CHECK(response.status == HTTP_STATUS_OK);
 
@@ -811,7 +810,7 @@ TEST_CASE("JsonWrappedEndpointFunction")
       const auto get_caller = create_simple_request("/get_caller", pack_type);
       const auto serialized_call = get_caller.build_request();
 
-      auto rpc_ctx = enclave::make_rpc_context(user_session, serialized_call);
+      auto rpc_ctx = ccf::make_rpc_context(user_session, serialized_call);
       auto response = parse_response(frontend.process(rpc_ctx).value());
       CHECK(response.status == HTTP_STATUS_OK);
 
@@ -825,7 +824,7 @@ TEST_CASE("JsonWrappedEndpointFunction")
     auto dont_fail = create_simple_request("/failable");
     const auto serialized_call = dont_fail.build_request();
 
-    auto rpc_ctx = enclave::make_rpc_context(user_session, serialized_call);
+    auto rpc_ctx = ccf::make_rpc_context(user_session, serialized_call);
     auto response = parse_response(frontend.process(rpc_ctx).value());
     CHECK(response.status == HTTP_STATUS_OK);
   }
@@ -846,7 +845,7 @@ TEST_CASE("JsonWrappedEndpointFunction")
       fail.set_body(serialized_body.data(), serialized_body.size());
       const auto serialized_call = fail.build_request();
 
-      auto rpc_ctx = enclave::make_rpc_context(user_session, serialized_call);
+      auto rpc_ctx = ccf::make_rpc_context(user_session, serialized_call);
       auto response = parse_response(frontend.process(rpc_ctx).value());
       CHECK(response.status == err);
       CHECK(
@@ -877,7 +876,7 @@ TEST_CASE("Restricted verbs")
     {
       http::Request get("get_only", verb);
       const auto serialized_get = get.build_request();
-      auto rpc_ctx = enclave::make_rpc_context(user_session, serialized_get);
+      auto rpc_ctx = ccf::make_rpc_context(user_session, serialized_get);
       const auto serialized_response = frontend.process(rpc_ctx).value();
       const auto response = parse_response(serialized_response);
       if (verb == HTTP_GET)
@@ -897,7 +896,7 @@ TEST_CASE("Restricted verbs")
     {
       http::Request post("post_only", verb);
       const auto serialized_post = post.build_request();
-      auto rpc_ctx = enclave::make_rpc_context(user_session, serialized_post);
+      auto rpc_ctx = ccf::make_rpc_context(user_session, serialized_post);
       const auto serialized_response = frontend.process(rpc_ctx).value();
       const auto response = parse_response(serialized_response);
       if (verb == HTTP_POST)
@@ -918,7 +917,7 @@ TEST_CASE("Restricted verbs")
       http::Request put_or_delete("put_or_delete", verb);
       const auto serialized_put_or_delete = put_or_delete.build_request();
       auto rpc_ctx =
-        enclave::make_rpc_context(user_session, serialized_put_or_delete);
+        ccf::make_rpc_context(user_session, serialized_put_or_delete);
       const auto serialized_response = frontend.process(rpc_ctx).value();
       const auto response = parse_response(serialized_response);
       if (verb == HTTP_PUT || verb == HTTP_DELETE)
@@ -984,8 +983,7 @@ TEST_CASE("Explicit commitability")
       request.set_body(&serialized_body);
 
       const auto serialized_request = request.build_request();
-      auto rpc_ctx =
-        enclave::make_rpc_context(user_session, serialized_request);
+      auto rpc_ctx = ccf::make_rpc_context(user_session, serialized_request);
       const auto serialized_response = frontend.process(rpc_ctx).value();
       const auto response = parse_response(serialized_response);
 
@@ -1020,8 +1018,7 @@ TEST_CASE("Explicit commitability")
         request.set_body(&serialized_body);
 
         const auto serialized_request = request.build_request();
-        auto rpc_ctx =
-          enclave::make_rpc_context(user_session, serialized_request);
+        auto rpc_ctx = ccf::make_rpc_context(user_session, serialized_request);
         const auto serialized_response = frontend.process(rpc_ctx).value();
         const auto response = parse_response(serialized_response);
 
@@ -1054,7 +1051,7 @@ TEST_CASE("Alternative endpoints")
     auto command = create_simple_request("/command");
     const auto serialized_command = command.build_request();
 
-    auto rpc_ctx = enclave::make_rpc_context(user_session, serialized_command);
+    auto rpc_ctx = ccf::make_rpc_context(user_session, serialized_command);
     auto response = parse_response(frontend.process(rpc_ctx).value());
     CHECK(response.status == HTTP_STATUS_OK);
   }
@@ -1064,8 +1061,7 @@ TEST_CASE("Alternative endpoints")
     http::Request read_only("read_only", verb);
     const auto serialized_read_only = read_only.build_request();
 
-    auto rpc_ctx =
-      enclave::make_rpc_context(user_session, serialized_read_only);
+    auto rpc_ctx = ccf::make_rpc_context(user_session, serialized_read_only);
     auto response = parse_response(frontend.process(rpc_ctx).value());
     CHECK(response.status == HTTP_STATUS_OK);
   }
@@ -1081,7 +1077,7 @@ TEST_CASE("Templated paths")
     auto request = create_simple_request("/fin/fang/foom");
     const auto serialized_request = request.build_request();
 
-    auto rpc_ctx = enclave::make_rpc_context(user_session, serialized_request);
+    auto rpc_ctx = ccf::make_rpc_context(user_session, serialized_request);
     auto response = parse_response(frontend.process(rpc_ctx).value());
     CHECK(response.status == HTTP_STATUS_OK);
 
@@ -1100,7 +1096,7 @@ TEST_CASE("Templated paths")
     auto request = create_simple_request("/users/1/address");
     const auto serialized_request = request.build_request();
 
-    auto rpc_ctx = enclave::make_rpc_context(user_session, serialized_request);
+    auto rpc_ctx = ccf::make_rpc_context(user_session, serialized_request);
     auto response = parse_response(frontend.process(rpc_ctx).value());
     CHECK(response.status == HTTP_STATUS_OK);
 
@@ -1127,8 +1123,7 @@ TEST_CASE("Signed read requests can be executed on backup")
 
   auto signed_call = create_signed_request(user_caller);
   auto serialized_signed_call = signed_call.build_request();
-  auto rpc_ctx =
-    enclave::make_rpc_context(user_session, serialized_signed_call);
+  auto rpc_ctx = ccf::make_rpc_context(user_session, serialized_signed_call);
   auto response = parse_response(frontend.process(rpc_ctx).value());
   CHECK(response.status == HTTP_STATUS_OK);
 }
@@ -1147,8 +1142,8 @@ TEST_CASE("Forwarding" * doctest::test_suite("forwarding"))
   network_primary.tables->set_consensus(primary_consensus);
 
   auto channel_stub = std::make_shared<ChannelStubProxy>();
-  auto rpc_responder = std::weak_ptr<enclave::AbstractRPCResponder>();
-  auto rpc_map = std::weak_ptr<enclave::RPCMap>();
+  auto rpc_responder = std::weak_ptr<ccf::AbstractRPCResponder>();
+  auto rpc_map = std::weak_ptr<ccf::RPCMap>();
   auto backup_forwarder = std::make_shared<Forwarder<ChannelStubProxy>>(
     rpc_responder, channel_stub, rpc_map, ConsensusType::CFT);
   auto backup_consensus = std::make_shared<kv::test::BackupStubConsensus>();
@@ -1157,9 +1152,8 @@ TEST_CASE("Forwarding" * doctest::test_suite("forwarding"))
   auto simple_call = create_simple_request();
   auto serialized_call = simple_call.build_request();
 
-  auto backup_ctx =
-    enclave::make_rpc_context(backup_user_session, serialized_call);
-  auto ctx = enclave::make_rpc_context(user_session, serialized_call);
+  auto backup_ctx = ccf::make_rpc_context(backup_user_session, serialized_call);
+  auto ctx = ccf::make_rpc_context(user_session, serialized_call);
 
   {
     INFO("Backup frontend without forwarder does not forward");
@@ -1174,7 +1168,7 @@ TEST_CASE("Forwarding" * doctest::test_suite("forwarding"))
   }
 
   user_frontend_backup.set_cmd_forwarder(backup_forwarder);
-  backup_ctx->session->is_forwarding = false;
+  backup_ctx->get_session_context()->is_forwarding = false;
 
   {
     INFO("Read command is not forwarded to primary");
@@ -1262,7 +1256,7 @@ TEST_CASE("Forwarding" * doctest::test_suite("forwarding"))
     auto signed_call = create_signed_request(user_caller);
     auto serialized_signed_call = signed_call.build_request();
     auto signed_ctx =
-      enclave::make_rpc_context(user_session, serialized_signed_call);
+      ccf::make_rpc_context(user_session, serialized_signed_call);
     const auto r = user_frontend_backup.process(signed_ctx);
     REQUIRE(!r.has_value());
     REQUIRE(channel_stub->size() == 1);
@@ -1278,7 +1272,7 @@ TEST_CASE("Forwarding" * doctest::test_suite("forwarding"))
 
   // On a session that was previously forwarded, and is now primary,
   // commands should still succeed
-  ctx->session->is_forwarding = true;
+  ctx->get_session_context()->is_forwarding = true;
   {
     INFO("Write command primary on a forwarded session succeeds");
     REQUIRE(channel_stub->is_empty());
@@ -1309,8 +1303,8 @@ TEST_CASE("Nodefrontend forwarding" * doctest::test_suite("forwarding"))
   auto primary_consensus = std::make_shared<kv::test::PrimaryStubConsensus>();
   network_primary.tables->set_consensus(primary_consensus);
 
-  auto rpc_responder = std::weak_ptr<enclave::AbstractRPCResponder>();
-  auto rpc_map = std::weak_ptr<enclave::RPCMap>();
+  auto rpc_responder = std::weak_ptr<ccf::AbstractRPCResponder>();
+  auto rpc_map = std::weak_ptr<ccf::RPCMap>();
   auto backup_forwarder = std::make_shared<Forwarder<ChannelStubProxy>>(
     rpc_responder, channel_stub, rpc_map, ConsensusType::CFT);
   node_frontend_backup.set_cmd_forwarder(backup_forwarder);
@@ -1320,9 +1314,9 @@ TEST_CASE("Nodefrontend forwarding" * doctest::test_suite("forwarding"))
   auto write_req = create_simple_request();
   auto serialized_call = write_req.build_request();
 
-  auto node_session = std::make_shared<enclave::SessionContext>(
-    enclave::InvalidSessionId, node_caller.raw());
-  auto ctx = enclave::make_rpc_context(node_session, serialized_call);
+  auto node_session = std::make_shared<ccf::SessionContext>(
+    ccf::InvalidSessionId, node_caller.raw());
+  auto ctx = ccf::make_rpc_context(node_session, serialized_call);
   const auto r = node_frontend_backup.process(ctx);
   REQUIRE(!r.has_value());
   REQUIRE(channel_stub->size() == 1);
@@ -1355,8 +1349,8 @@ TEST_CASE("Userfrontend forwarding" * doctest::test_suite("forwarding"))
   auto primary_consensus = std::make_shared<kv::test::PrimaryStubConsensus>();
   network_primary.tables->set_consensus(primary_consensus);
 
-  auto rpc_responder = std::weak_ptr<enclave::AbstractRPCResponder>();
-  auto rpc_map = std::weak_ptr<enclave::RPCMap>();
+  auto rpc_responder = std::weak_ptr<ccf::AbstractRPCResponder>();
+  auto rpc_map = std::weak_ptr<ccf::RPCMap>();
   auto backup_forwarder = std::make_shared<Forwarder<ChannelStubProxy>>(
     rpc_responder, channel_stub, rpc_map, ConsensusType::CFT);
   user_frontend_backup.set_cmd_forwarder(backup_forwarder);
@@ -1366,7 +1360,7 @@ TEST_CASE("Userfrontend forwarding" * doctest::test_suite("forwarding"))
   auto write_req = create_simple_request();
   auto serialized_call = write_req.build_request();
 
-  auto ctx = enclave::make_rpc_context(user_session, serialized_call);
+  auto ctx = ccf::make_rpc_context(user_session, serialized_call);
   const auto r = user_frontend_backup.process(ctx);
   REQUIRE(!r.has_value());
   REQUIRE(channel_stub->size() == 1);
@@ -1403,8 +1397,8 @@ TEST_CASE("Memberfrontend forwarding" * doctest::test_suite("forwarding"))
   auto primary_consensus = std::make_shared<kv::test::PrimaryStubConsensus>();
   network_primary.tables->set_consensus(primary_consensus);
 
-  auto rpc_responder = std::weak_ptr<enclave::AbstractRPCResponder>();
-  auto rpc_map = std::weak_ptr<enclave::RPCMap>();
+  auto rpc_responder = std::weak_ptr<ccf::AbstractRPCResponder>();
+  auto rpc_map = std::weak_ptr<ccf::RPCMap>();
   auto backup_forwarder = std::make_shared<Forwarder<ChannelStubProxy>>(
     rpc_responder, channel_stub, rpc_map, ConsensusType::CFT);
   member_frontend_backup.set_cmd_forwarder(backup_forwarder);
@@ -1414,7 +1408,7 @@ TEST_CASE("Memberfrontend forwarding" * doctest::test_suite("forwarding"))
   auto write_req = create_simple_request();
   auto serialized_call = write_req.build_request();
 
-  auto ctx = enclave::make_rpc_context(member_session, serialized_call);
+  auto ctx = ccf::make_rpc_context(member_session, serialized_call);
   const auto r = member_frontend_backup.process(ctx);
   REQUIRE(!r.has_value());
   REQUIRE(channel_stub->size() == 1);
@@ -1494,7 +1488,7 @@ TEST_CASE("Retry on conflict")
     size_t retry_count = ccf_max_attempts - 1;
     req.set_header("test-retry-count", fmt::format("{}", retry_count));
     auto serialized_call = req.build_request();
-    auto rpc_ctx = enclave::make_rpc_context(user_session, serialized_call);
+    auto rpc_ctx = ccf::make_rpc_context(user_session, serialized_call);
 
     auto response = parse_response(frontend.process(rpc_ctx).value());
     CHECK(response.status == HTTP_STATUS_OK);
@@ -1510,7 +1504,7 @@ TEST_CASE("Retry on conflict")
     size_t retry_count = ccf_max_attempts + 1;
     req.set_header("test-retry-count", fmt::format("{}", retry_count));
     auto serialized_call = req.build_request();
-    auto rpc_ctx = enclave::make_rpc_context(user_session, serialized_call);
+    auto rpc_ctx = ccf::make_rpc_context(user_session, serialized_call);
 
     auto response = parse_response(frontend.process(rpc_ctx).value());
     CHECK(response.status == HTTP_STATUS_SERVICE_UNAVAILABLE);